LUADIR needs PREFIX
Sorry for missing this in review of #38.

Also fixed double-slashes resulting from the `$(DESTDIR)/` pattern.
Since the various sub-paths start with `$(PREFIX)` which has a
leading slash, we don't need another one separating DESTDIR from the
rest of the path.

Another thing I missed in review is that we should install executable
assets 0555 rather than 0755.